[refactor] Shove pretty much everything into `wordpresses.py`
