+++ ansible/roles/wordpress-namespace/tasks/wp-cron.yml
@@ -23,7 +23,7 @@
             serviceAccountName: wp-cron
             containers:
               - name: wp-cron
-                image: "quay-its.epfl.ch/svc0041/wp-cron:2025-015"
+                image: "quay-its.epfl.ch/svc0041/wp-cron:2025-017"
                 env:
                   - name: K8S_NAMESPACE
                     valueFrom:
